feat: create join.html and add two methods in HelloController.java that will have a Get and Post method that will take in the value of the input and place that into the h1 tag in join.html

+++ src/main/java/com/codeup/adlister/controllers/PostController.java
@@ -13,7 +13,7 @@ public class PostController {
         return "posts index page";
     }
     @GetMapping("/posts/{id}") @ResponseBody
-    public String IndividualPost(@PathVariable int id) {
+    public String IndividualPost(@PathVariable long id) {
         return "view user with id of: " +id+ "'s post";
     }
     @GetMapping("/posts/create") @ResponseBody
